+++ src/main/java/ee/ut/math/vl/servlets/kasutaja/KasutajaController.java
@@ -57,7 +57,7 @@ public class KasutajaController extends HttpServlet {
 
     private void  replyWithKasutaja(HttpServletResponse resp,
                                          String idString) throws SQLException, Exception {
-        int id = Integer.parseInt(idString);
+        long id = Long.parseLong(idString);
 
 
         Kasutaja kasutaja = datastore.findKasutajaById(id);
